move data to Overview.test.data.js and import to Overview.test.js
